refactor: orderbook bench helper
